pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "umafall-mtl"
version = "1.0.0"
description = "Multi-task learning pipeline for UMAFall dataset"
readme = "docs/README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24.3",
    "pandas>=2.0.3",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0.1",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.4.1",
    "pre-commit>=3.3.3",
]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true